+++ include/pile/utils/common.hpp
@@ -7,9 +7,12 @@
 #include <cstddef>
 #include <cxxopts.hpp>
 #include <fstream>
+#include <map>
 #include <numeric>
 #include <pile/utils/core.hpp>
 #include <pile/utils/stack.hpp>
 #include <pile/utils/utils.hpp>
 #include <ranges>
 #include <regex>
+#include <set>
+#include <variant>